+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/access/rateLimit/index.js
@@ -1,0 +1,32 @@
+import angular from 'angular';
+import ngOvhTelecomUniverseComponents from '@ovh-ux/ng-ovh-telecom-universe-components';
+import '@uirouter/angularjs';
+import '@ovh-ux/ng-translate-async-loader';
+import 'angular-translate';
+import 'ovh-api-services';
+
+import controller from './ratelimit.controller';
+import template from './ratelimit.html';
+
+const moduleName = 'XdslAccessLnsRateLimit';
+
+angular
+  .module(moduleName, [
+    ngOvhTelecomUniverseComponents,
+    'ngTranslateAsyncLoader',
+    'pascalprecht.translate',
+    'ovh-api-services',
+    'ui.router',
+  ])
+  .controller('XdslAccessLnsRateLimitCtrl', controller)
+  .run(/* @ngTranslationsInject:json ./translations */)
+  .run(
+    /* @ngInject */ ($templateCache) => {
+      $templateCache.put(
+        'app/telecom/pack/xdsl/access/rateLimit/xdsl-access-lns-ratelimit.html',
+        template,
+      );
+    },
+  );
+
+export default moduleName;
