+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/index.js
@@ -10,7 +10,6 @@ import missingRio from './missingRio';
 
 import component from './pack-xdsl.component';
 import routing from './pack-xdsl.routing';
-import templates from './pack-xdsl.templates';
 
 const moduleName = 'ovhManagerTelecomPackXdsl';
 
@@ -26,7 +25,6 @@ angular
   ])
   .component('packXdsl', component)
   .config(routing)
-  .run(templates)
   .run(/* @ngTranslationsInject:json ./translations */);
 
 export default moduleName;
